# server/swagger.yaml
openapi: 3.0.0
info:
  title: FloNeo LCNC Platform API
  description: RESTful API for the FloNeo Low-Code No-Code Platform
  version: 1.0.0
  contact:
    name: FloNeo Team
    email: support@floneo.com
servers:
  - url: http://localhost:5000
    description: Development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/notifications/read:
    get:
      summary: Mark notifications as read
      description: Mark all unread notifications for the authenticated user as read and return them
      tags:
        - Notifications
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Notifications marked as read successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "2 notifications marked as read"
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        type:
                          type: string
                          example: "invite"
                        message:
                          type: string
                          example: "You have been invited to a project"
                        read:
                          type: boolean
                          example: true
                        createdAt:
                          type: string
                          format: date-time
        '401':
          description: Authentication required
  /api/notifications:
    get:
      summary: Get user notifications
      description: Retrieve notifications for the authenticated user with pagination
      tags:
        - Notifications
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of notifications per page
          schema:
            type: integer
            default: 10
        - name: unreadOnly
          in: query
          description: Filter to show only unread notifications
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Notifications retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        type:
                          type: string
                        message:
                          type: string
                        read:
                          type: boolean
                        createdAt:
                          type: string
                          format: date-time
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      pages:
                        type: integer
        '401':
          description: Authentication required
...